Fix postinstall when nativescript is not in the PATH
In case `nativescript` is not in the PATH variable, trying to execute `nativescript --version` will fail and the rest of the postinstall will not be executed.
For example on many test environments, where different CLI versions are tested, CLI is installed in a directory and executed directly from there.
Also this breaks adding the plugin with AppBuilder CLI (and other AppBuilder clients).

So try-catch the execution.