Fix usage of plugin with npm 2
Whenever you try using the plugin with `npm 2`, it's postinstall script will fail, as it requires `fs-extra` module. However it's not been added as dependency of the plugin, so it does not exist in node_modules and `require` fails.
When npm 3 or later is used, the `fs-extra` plugin exists in the root level of `node_modules` as it is a dependency of `fs-promise`. npm 3 and later tries to flatten the dependencies tree, so `fs-extra` is moved in the root level and the `require` in the post-install script works.

In order to fix this, add the `fs-extra` as dependency of the plugin.

NOTE: Maybe it's worth checking why we need fs-extra, Node.js `fs` could do the work, but I leave it for further discussions.