Mark init and getRemoteConfig for profiling
These sometimes complete async or are called in the application launched event and the times spend inside are otherwise hard to track. Wrapping these in profile will show them when running the timeline profiling. init on my android phone spends about 150ms, that's fine but I'd like to be able to keep an eye on it.

I did the changes in the GitHub UI, so please test it before merge.