Add improved iOS logging
This addition adds more insight into iOS crashes, especially for failures in JS that result in the crash.
Using this logging we were able to identify this issue: https://github.com/NativeScript/NativeScript/pull/5475 which otherwise would have been impossible to trace, as it had a 5% repro among users but never occurred internally.