Minimum ios version
Hi @EddyVerbruggen, 
The work you've made on this project, is simply amazing.  Thanks a lot.
I am a new a newbee in nativescript, but I have struggle with a compilation issue.
If you think this additional line can be useful. otherwise, no problem and thanks again for your great job.