Adding information regarding access rules.
If the listeners are used, they are indicated to be added after
authentication to firebase, if the rules prevent access to data without
authentication.

TODO: Improve handling of rule-related errors by giving more details of
what caused the error if possible.

TODO: Check if there is a way to avoid such an access problem.